+++ zbar/config.c
@@ -36,14 +36,14 @@ int zbar_parse_config (const char *cfgstr,
                        int *val)
 {
     const char *dot, *eq;
-    int len;
+    long len;
     char negate;
     if(!cfgstr)
         return(1);
 
     dot = strchr(cfgstr, '.');
     if(dot) {
-        int len = dot - cfgstr;
+        long len = dot - cfgstr;
         if(!len || (len == 1 && !strncmp(cfgstr, "*", len)))
             *sym = 0;
         else if(len < 2)
