+++ zbar/image.c
@@ -241,7 +241,7 @@ typedef struct zimg_hdr_s {
 int zbar_image_write (const zbar_image_t *img,
                       const char *filebase)
 {
-    int len = strlen(filebase) + 16;
+    long len = strlen(filebase) + 16;
     char *filename = malloc(len);
     int n = 0, rc = 0;
     FILE *f;
